Alex: Completed the implementation of MetaData
This includes the creation  of the meta data class and a new wrapper method called HashID that wraps a hash and converts strings to the hash number implicitely. The elements of the MetaData can be iterated through using a range  for loop or an explicit call to begin and end.